improved ~purge command... maybe

+++ commands/ADMIN.py
@@ -59,22 +59,23 @@ class Purge:
                 count -= 1
         while len(delM):
             try:
-                try:
+                if hasattr(channel, "delete_messages"):
                     await channel.delete_messages(delM[:100])
                     deleted += min(len(delM), 100)
                     for i in range(min(len(delM), 100)):
                         delM.popleft()
-                except AttributeError:
+                else:
                     _vars.logDelete(delM[0].id)
                     await delM[0].delete()
                     deleted += 1
                     delM.popleft()
             except:
                 print(traceback.format_exc())
-                m = delM.popleft()
-                _vars.logDelete(m.id)
-                await m.delete()
-                deleted += 1
+                for i in range(min(5, len(delM))):
+                    m = delM.popleft()
+                    _vars.logDelete(m.id)
+                    await m.delete()
+                    deleted += 1
         if not "h" in flags:
             return (
                 "```css\nDeleted " + uniStr(deleted)
